feat(tempo): add accessKey.verifyHash and use it in chainConfig.verifyHash
Amp-Thread-ID: https://ampcode.com/threads/T-019e7a56-c1d1-74f3-b825-3e10c7334663
