+++ .changeset/tempo-admin-access-keys.md
@@ -2,4 +2,4 @@
 "viem": minor
 ---
 
-**`viem/tempo`:** Added admin access key actions, and `signatureVerifier.verifyKeychain` & `signatureVerifier.verifyKeychainAdmin` actions.
+**`viem/tempo`:** Added admin access key actions, and an `accessKey.verifyHash` action.
