--- site/pages/tempo/actions/signatureVerifier.verifyKeychain.mdx
@@ -1,60 +1,0 @@
-import ReadParameters from '../../../snippets/tempo/read-parameters.mdx'
-
-# `signatureVerifier.verifyKeychain`
-
-Verifies that a keychain signature was produced by an active access key for the expected account.
-
-Returns `false` for account mismatches, unknown, revoked, or expired access keys.
-
-[TIP-1049](https://tips.sh/1049)
-
-## Usage
-
-:::code-group
-
-```ts twoslash [example.ts]
-import { client } from './viem.config'
-
-const valid = await client.signatureVerifier.verifyKeychain({
-  account: '0x742d35Cc6634C0532925a3b844Bc9e7595f0bEbb',
-  hash: '0xdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeef',
-  signature: '0x...',
-})
-
-console.log('Valid:', valid)
-// @log: Valid: true
-```
-
-```ts twoslash [viem.config.ts] filename="viem.config.ts"
-// [!include ~/snippets/tempo/viem.config.ts:setup]
-```
-
-:::
-
-## Return Type
-
-```ts
-type ReturnType = boolean
-```
-
-## Parameters
-
-### account
-
-- **Type:** `Address`
-
-Account address the signature is expected to belong to.
-
-### hash
-
-- **Type:** `Hex`
-
-Original message hash that was signed.
-
-### signature
-
-- **Type:** `Hex`
-
-Keychain signature envelope (V2). Legacy keychain signatures are rejected.
-
-<ReadParameters />
