+++ src/tempo/actions/index.ts
@@ -9,7 +9,6 @@ export * as nonce from './nonce.js'
 export * as policy from './policy.js'
 export * as receivePolicy from './receivePolicy.js'
 export * as reward from './reward.js'
-export * as signatureVerifier from './signatureVerifier.js'
 export * as simulate from './simulate.js'
 export * as token from './token.js'
 export * as validator from './validator.js'
